+++ app/app.go
@@ -472,6 +472,7 @@ func NewEthermintApp(
 		tracer,
 		evmSs, nil,
 		allKeys,
+		nil,
 	)
 
 	// Create IBC Keeper